refactor: split spatialdata converter into modular architecture
- Split monolithic 747-line spatialdata_converter.py into focused modules:
  * base_spatialdata_converter.py - shared functionality
  * spatialdata_2d_converter.py - 2D slice handling
  * spatialdata_3d_converter.py - 3D volume handling
  * converter.py - factory pattern for automatic selection

- Fix critical bugs:
  * Replace reader.get_dimensions() with reader.get_essential_metadata().dimensions
  * Fix hard-coded version string to import dynamically from __init__.py
  * Add input validation for pixel_size_um > 0 and non-empty dataset_id
  * Fix performance regression by using total_intensity for 3D average spectrum

- Improve scientific accuracy:
  * Change column name from "mass" to "mz" for mass spectrometry data
  * Update docstrings to reflect m/z terminology

- Clean package structure:
  * Move factory class to dedicated converter.py file
  * Clean __init__.py to handle only imports and registration
  * Update VS Code settings to hide development artifacts
  * Add *.ipynb to .gitignore

- Maintain backward compatibility:
  * Same import path: from msiconvert.converters.spatialdata import SpatialDataConverter
  * Identical API and functionality
  * All existing tests should work without modification

This refactoring improves maintainability while preserving all functionality
and fixing several critical issues that were causing conversion failures.
