+++ msiconvert/converters/__init__.py
@@ -1,9 +1,9 @@
 import logging
 
 try:
-    from . import spatialdata_converter
+    from . import spatialdata
 
-    logging.debug("Successfully imported spatialdata_converter")
+    logging.debug("Successfully imported spatialdata package")
 except (ImportError, NotImplementedError) as e:
     # Skip if spatialdata dependencies not available or incompatible
     logging.error(f"SpatialData converter not available due to dependency issues: {e}")
@@ -11,7 +11,7 @@ except (ImportError, NotImplementedError) as e:
 
     logging.error(f"Full traceback: {traceback.format_exc()}")
 except Exception as e:
-    logging.error(f"Unexpected error importing spatialdata_converter: {e}")
+    logging.error(f"Unexpected error importing spatialdata package: {e}")
     import traceback
 
     logging.error(f"Full traceback: {traceback.format_exc()}")
