+++ msiconvert/converters/spatialdata/__init__.py
@@ -1,0 +1,7 @@
+# msiconvert/converters/spatialdata/__init__.py
+
+from .converter import SpatialDataConverter
+from .spatialdata_2d_converter import SpatialData2DConverter
+from .spatialdata_3d_converter import SpatialData3DConverter
+
+__all__ = ["SpatialDataConverter", "SpatialData2DConverter", "SpatialData3DConverter"]
